+++ pages/_app.tsx
@@ -2,14 +2,16 @@ import '../styles/globals.css';
 import type { AppProps } from 'next/app';
 import GlobalStyle from '../styles/GlobalStyle';
 import Header from '../components/Header';
+import { ThemeProvider } from 'styled-components';
+import { theme } from '../styles/theme';
 
 function MyApp({ Component, pageProps }: AppProps) {
   return (
-    <>
+    <ThemeProvider theme={theme}>
       <GlobalStyle />
       <Header />
       <Component {...pageProps} />
-    </>
+    </ThemeProvider>
   );
 }
